iv.sensemakr: Sensitivity Analysis Tools for Instrumental Variable Estimates

Implements a suite of sensitivity analysis tools for instrumental variable estimates, as described in Cinelli and Hazlett (2025).

Details

The main functions of the package are iv_fit, which fits an IV regression model using the Anderson-Rubin approach, and sensemakr.iv_fit, which computes sensitivity analysis results for objects of class iv_fit.

References

Cinelli, C. and Hazlett, C. (2025), "An Omitted Variable Bias Framework for Sensitivity Analysis of Instrumental Variables." Biometrika. doi:10.1093/biomet/asaf004

See also

Author

Maintainer: Carlos Cinelli carloscinelli@hotmail.com

Authors:

Chad Hazlett